+++ src/brekekejs/index.d.ts
@@ -204,6 +204,10 @@ export type PbxPal = {
   sendDTMF(p: PbxSendDtmfParam, resolve: () => void, reject: ErrorHandler): void
 }
 
+export type PbxResourceLine = {
+  key: string
+  value: string
+}
 export type PbxCustomPage = {
   id: string
   url: string
@@ -240,6 +244,7 @@ export type PbxGetProductInfoRes = {
   'webrtcclient.dtmfSendMode': string
   'webphone.phonebook.personal.editable': string
   'webphone.recents.max': string
+  'webphone.resource-line': string
   version: string
 }
 export type PbxGetProductInfoParam = {
@@ -405,7 +410,7 @@ export type MakeCallFn = (
   options?: object,
   videoEnabled?: boolean,
   videoOptions?: object,
-  exInfo?: object,
+  exInfo?: string,
 ) => void
 
 export type VideoOptions = {
@@ -486,6 +491,9 @@ export type Session = {
     }
     direction: 'outgoing' | 'incoming'
     terminate(): void
+    _request?: {
+      extraHeaders?: string[]
+    }
   }
   withVideo: boolean
   remoteWithVideo: boolean