+++ src/stores/addCallHistory.ts
@@ -96,6 +96,12 @@ export const addCallHistory = async (
   const created = moment().format('HH:mm - MMM D')
   const answeredBy = getUserInfoFromReasons(isTypeCall ? ms : completedBy)
   const reason = getReasonCancelCall(answeredBy)
+  const line = (isTypeCall && c?.line) || undefined
+  // with incoming: If the string includes /, you store only aaa to the log and ignore / and the following string.
+  const lineValue = line?.split('/')?.[0]?.trim() || line?.trim() || undefined
+  const lineLabel = lineValue
+    ? getAuthStore().resourceLines?.find(item => item.value === lineValue)?.key
+    : undefined
 
   const info = isTypeCall
     ? {
@@ -109,6 +115,8 @@ export const addCallHistory = async (
         isAboutToHangup: c.isAboutToHangup,
         reason,
         answeredBy,
+        lineValue,
+        lineLabel,
       }
     : {
         id,
@@ -165,6 +173,8 @@ export type CallHistoryInfo = {
   isAboutToHangup: boolean
   reason?: string
   answeredBy?: { name: string; phoneNumber: string }
+  lineLabel?: string
+  lineValue?: string
 }
 
 const addToCallLog = async (c: CallHistoryInfo) => {